# wallet/docs/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /api/wallet/withdraw/{account}:
    post:
      tags:
      - wallet-controller
      operationId: withdraw
      parameters:
      - name: account
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: request
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/WithdrawRequest'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/OperationResponse'
  /api/wallet/find-account:
    post:
      tags:
      - wallet-controller
      operationId: findAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginAccountRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BalanceResponse'
  /api/wallet/deposit/{account}:
    post:
      tags:
      - wallet-controller
      operationId: deposit
      parameters:
      - name: account
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: request
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/DepositRequest'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/OperationResponse'
  /api/wallet/account:
    post:
      tags:
      - wallet-controller
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewAccountRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BalanceResponse'
  /api/wallet/operations/{account}:
    get:
      tags:
      - wallet-controller
      operationId: listOperations
      parameters:
      - name: account
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: take
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/OperationsResponse'
  /api/wallet/balance/{account}:
    get:
      tags:
      - wallet-controller
      operationId: balance
      parameters:
      - name: account
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BalanceResponse'
components:
  schemas:
    WithdrawRequest:
      type: object
      properties:
        amount:
          type: integer
          format: int64
        reference:
          type: string
    OperationResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
    LoginAccountRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
    BalanceResponse:
      type: object
      properties:
        accountId:
          type: integer
          format: int64
        amount:
          type: number
    DepositRequest:
      type: object
      properties:
        amount:
          type: number
          format: double
        reference:
          type: string
    NewAccountRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
    Operation:
      type: object
      properties:
        type:
          type: string
        amount:
          type: number
        reference:
          type: string
    OperationsResponse:
      type: object
      properties:
        operations:
          type: array
          items:
            $ref: '#/components/schemas/Operation'
        hasMore:
          type: boolean